+++ sandbox/README.md
@@ -1,0 +1,3 @@
+# Sandbox
+
+Sandbox contains static data that is used to simulate an investing environment. This can be used to present openx in conferences or to your friends without having to worry about creating projects and populating them with stablecoin and XLM each time. The sandbox requires about 10 minutes to populate and can be run with the help of the `-t` flag.
