+++ sandbox/data/10mw.json
@@ -160,7 +160,7 @@
       "value": "https://images.openx.solar/OpenSolarProjects/7_YaleMIT/10.jpg"
     }, {
       "type": "text",
-      "value": "IoT meters include a Schneider Commbox and the Yale Open Powermer devices equiped with WiFi and Zigbee transfer protocol. A Rasberry Pi acts as a light clients"
+      "value": "IoT meters include a Schneider Commbox and the Yale Open Powermer devices equipped with WiFi and Zigbee transfer protocol. A Rasberry Pi acts as a light clients"
     }]
   }, {
     "width": 4,
