+++ sandbox/data/1kw.json
@@ -160,7 +160,7 @@
       "value": "https://images.openx.solar/demo/pasto-page-it-infrastructure.png"
     }, {
       "type": "text",
-      "value": "IoT meters include a Schneider Commbox and the Yale Open Powermer devices equiped with WiFi and Zigbee transfer protocol. A Rasberry Pi acts as a light clients"
+      "value": "IoT meters include a Schneider Commbox and the Yale Open Powermer devices equipped with WiFi and Zigbee transfer protocol. A Rasberry Pi acts as a light clients"
     }]
   }, {
     "width": 4,
