--- stellar.toml
@@ -1,10 +1,0 @@
-# Stellar.toml is equired to be hosted on the public server we run the frontend
-# on to advertise to peers what kind of assets we support
-[[CURRENCIES]]
-code="INVToken"
-issuer="" # fill in public key once final
-display_decimals=7
-name="YOL Investor Token"
-desc="Investor Tokens that are handed out for every USD an investor invests"
-conditions="" # Insert conditions here
-image="https://upload.wikimedia.org/wikipedia/commons/b/b2/Petro_%28cryptocurrency%29_logo.png" # Photo of petro, change this
